+++ src/components/Header.jsx
@@ -145,92 +145,4 @@ export default Header;
     export default Section;
     
     
-    
-
-
-    import React from 'react';
-
-const About = () => {
-  return (import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-
-const EmployeeList = () => {
-  const [employees, setEmployees] = useState([]);
-
-  useEffect(() => {
-    axios.get('http://localhost:8080/api/employees')
-      .then(res => setEmployees(res.data))
-      .catch(err => console.error(err));
-  }, []);
-
-  return (
-    <div>
-      <h2>Employees</h2>
-      <ul>
-        {employees.map(emp => (
-          <li key={emp.id}>{emp.name} - {emp.email}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-
-import React from 'react';
-import { Box, Typography, styled } from '@mui/material';
-
-const Component = styled(Box)({
-  width: '100%',
-  height: 580,
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  paddingTop: 47,
-  objectFit: 'cover',
-  backgroundRepeat: 'no-repeat',
-  backgroundSize: 'cover',
-});
-
-const Heading = styled(Typography)({
-  fontWeight: 600,
-  fontSize: 56,
-});
-
-const SubHeading = styled(Typography)({
-  fontSize: 28,
-  padding: '0 0 5px 0',
-});
-
-const linkStyle = {
-  color: '#2997ff',
-  padding: '0 30px',
-  fontSize: 21,
-  cursor: 'pointer',
-};
-
-const Section = ({ data = {} }) => {
-  const {
-    imageURL = '',
-    heading = '',
-    headColor = '#000',
-    subHeading = '',
-    subHeadColor = '#333',
-  } = data;
-
-  return (
-    <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-      <Heading style={{ color: headColor }}>{heading}</Heading>
-      <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-      <Typography>
-        <Typography component="span" style={linkStyle}>
-          Learn more &gt;
-        </Typography>
-        <Typography component="span" style={linkStyle}>
-          Buy &gt;
-        </Typography>
-      </Typography>
-    </Component>
-  );
-};
-
-export default Section;
+
