+++ src/components/Product.jsx
@@ -55,61 +55,50 @@ const Product = ({ data = {} }) => {
 
 export default Product;
 
+import React from 'react';
+import { Link } from 'react-router-dom';
+import './Navbar.css'; // optional for custom styling
 
-import { Box, Typography, styled } from '@mui/material';
+const Navbar = () => {
+  return (
+    <nav className="navbar">
+      <h2 className="logo">MyAppimport React, { useEffect, useState } from 'react';
+import axios from 'axios';
 
-// Styled Components efceceerf13f3fihewdfe;qwd  sD  Shfwehdwd;.wodqwdqdqdq1dxqd    qhd.whF,JFGJM
-const Component = styled(Box)({
-    width: '100%',
-    height: 580,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    paddingTop: 47,
-    objectFit: 'cover',
-    backgroundRepeat: 'no-repeat',
-    backgroundSize: 'cover'
-});
+const EmployeeList = () => {
+  const [employees, setEmployees] = useState([]);
 
-const Heading = styled(Typography)({
-    fontWeight: 600,
-    fontSize: 56
-});
+  useEffect(() => {
+    axios.get('http://localhost:8080/api/employees')
+      .then(res => setEmployees(res.data))
+      .catch(err => console.error(err));
+  }, []);
 
-const SubHeading = styled(Typography)({
-    fontSize: 28,
-    padding: '0 0 5px 0'
-});
-
-const linkStyle = {
-    color: '#2997ff',
-    padding: '0 30px',
-    fontSize: 21,
-    cursor: 'pointer'
+  return (
+    <div>
+      <h2>Employees</h2>
+      <ul>
+        {employees.map(emp => (
+          <li key={emp.id}>{emp.name} - {emp.email}</li>
+        ))}
+      </ul>
+    </div>
+  );
 };
 
-// Functional Component
-const Section = ({ data = {} }) => {
-    const {
-        imageURL = '',
-        heading = '',
-        headColor = '#000',
-        subHeading = '',
-        subHeadColor = '#333'
-    } = data;
-
-    return (
-        <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-            <Heading style={{ color: headColor }}>{heading}</Heading>
-            <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-            <Typography>
-                <Typography component="span" style={linkStyle}>Learn more &gt;</Typography>
-                <Typography component="span" style={linkStyle}>Buy &gt;</Typography>
-            </Typography>
-        </Component>
-    );
+export default EmployeeList;
+</h2>
+      <ul className="nav-links">
+        <li><Link to="/">Home</Link></li>
+        <li><Link to="/about">About</Link></li>
+        <li><Link to="/contact">Contact</Link></li>
+      </ul>
+    </nav>
+  );
 };
-<style>.navbar {
+<style>
+    
+.navbar {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -139,46 +128,7 @@ const Section = ({ data = {} }) => {
 }
 </style>
 
-export default Section;
-
-
-import React from 'react';
-
-const About = () => {
-  return (
-    <div stimport React, { useEffect, useState } from 'react';
-import axios from 'axios';
-
-const EmployeeList = () => {
-  const [employees, setEmployees] = useState([]);
-
-  useEffect(() => {
-    axios.get('http://localhost:8080/api/employees')
-      .then(res => setEmployees(res.data))
-      .catch(err => console.error(err));
-  }, []);
-
-  return (
-    <div>
-      <h2>Employees</h2>
-      <ul>
-        {employees.map(emp => (
-          <li key={emp.id}>{emp.name} - {emp.email}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-export default EmployeeList;
-yle={{ padding: '2rem', color: '#333' }}>
-      <h1>About Us</h1>
-      <p>This is the About page. Learn more about what we do!</p>
-    </div>
-  );
-};
-
-export default About;
+export default Navbar;
 import React from 'react';
 
 const About = () => {
@@ -191,7 +141,6 @@ const About = () => {
 };
 
 export default About;
-
 import { Box, Typography, styled } from '@mui/material';
 
 const Component = styled(Box)({
@@ -249,6 +198,7 @@ const Section = ({ data = {} }) => {
 };
 
 export default Section;
+
 import React from 'react';
 import { Box, Typography, styled } from '@mui/material';
 
@@ -283,27 +233,4 @@ const linkStyle = {
 
 const Section = ({ data = {} }) => {
   const {
-    imageURL = '',
-    heading = '',
-    headColor = '#000',
-    subHeading = '',
-    subHeadColor = '#333',
-  } = data;
-
-  return (
-    <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-      <Heading style={{ color: headColor }}>{heading}</Heading>
-      <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-      <Typography>
-        <Typography component="span" style={linkStyle}>
-          Learn more &gt;
-        </Typography>
-        <Typography component="span" style={linkStyle}>
-          Buy &gt;
-        </Typography>
-      </Typography>
-    </Component>
-  );
-};
-
-export default Section;
+    imageURL = ''
